requirements.txt
appdirs==1.4.0
configparser==3.5.0b2
docutils==0.12
ecdsa==0.13
GitPython==1.0.1
Jinja2==2.7.3
MarkupSafe==0.23
paramiko==1.15.2
PySide==1.2.2
PyYAML==3.11
Sphinx==1.3.1
SQLAlchemy==1.0.4
py==1.4.28
tox==2.0.1
